add entry update test.

+++ src/main/java/org/thiki/kanban/entry/EntriesService.java
@@ -1,11 +1,11 @@
 package org.thiki.kanban.entry;
 
-import java.util.List;
-
-import javax.annotation.Resource;
-
 import org.springframework.stereotype.Service;
 import org.thiki.kanban.foundation.common.Sequence;
+import org.thiki.kanban.foundation.exception.ResourceNotFoundException;
+
+import javax.annotation.Resource;
+import java.util.List;
 
 /**
  * Created by xubitao on 04/26/16.
@@ -47,6 +47,10 @@ public class EntriesService {
     }
 
     public int deleteById(String id) {
+        Entry entryToDelete = entriesPersistence.findById(id);
+        if (entryToDelete == null) {
+            throw new ResourceNotFoundException("entry[" + id + "] is not found.");
+        }
         return entriesPersistence.deleteById(id);
     }
 }
